src/main/python/ml-communication/requirements.txt
scikit-learn==0.23.2
flask==2.0.2
joblib==1.1.0
pandas==1.4.1
pillow==9.0.1
matplotlib==3.5.1